requirements.txt
resin-sdk==1.5.3
click==6.6
